move types into extension.ts
